requirements.txt
fastapi
uvicorn
requests
fitz
faiss-cpu
sentence-transformers
transformers
openai